requirements.txt.txt
Flask
pyodbc
google-generativeai
